rename tmp.py to test_scripts.py.
